requirements.txt
Flask==2.0.3
rembg==2.0.26
Pillow==8.4.0
gunicorn==20.1.0